backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Security - Fixed bcrypt version
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
cryptography==42.0.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilities
python-dotenv==1.0.0
redis==5.0.1

# Rate Limiting
slowapi==0.1.9

# Monitoring
prometheus-client==0.19.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
faker==22.0.0